Tagging version 8.2
git-svn-id: https://plugins.svn.wordpress.org/quick-featured-images/tags/8.2@967601 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -14,6 +14,23 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-// remove settings
-delete_option( 'quick-featured-images-settings' ); 
-delete_option( 'quick-featured-images-defaults' ); 
+if ( is_multisite() ) {
+
+	$sites = wp_get_sites();
+
+	if ( empty ( $sites ) ) return;
+
+	foreach ( $sites as $site ) {
+		// switch to next blog
+		switch_to_blog( $site[ 'blog_id' ] );
+		// remove settings
+		delete_option( 'quick-featured-images-settings' ); 
+		delete_option( 'quick-featured-images-defaults' );
+	}
+	// restore the current blog, after calling switch_to_blog()
+	restore_current_blog();
+} else {
+	// remove settings
+	delete_option( 'quick-featured-images-settings' ); 
+	delete_option( 'quick-featured-images-defaults' );
+}
